9_app_into_docker/requirements.txt
click==8.0.1
Flask==2.0.1
greenlet==1.1.0
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.18.0
python-json-logger==2.0.1
SQLAlchemy==1.4.20
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
